♻️ refactor(commit): extract commit scope helpers
